refactor(engine): split reason execution states
